requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.1
supabase>=2.3.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langchain-google-vertexai>=0.0.5
python-multipart>=0.0.6
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.5
pypdf>=3.17.0
requests>=2.31.0
